+++ composer.json
@@ -14,12 +14,12 @@
   ],
   "autoload": {
     "psr-4": {
-      "": "src/"
+      "Phrity\\O\\": "src/"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "": "tests/"
+      "Phrity\\O\\Test\\": "tests/"
     }
   },
   "require": {
