+++ src/Integer.php
