--- src/Phrity/O/Array/TypeInterface.php
@@ -1,18 +1,0 @@
-<?php
-
-/**
- * File for O\Array\TypeInterface interface.
- * @package Phrity > O
- */
-
-namespace Phrity\O\Array;
-
-/**
- * O\Array\TypeInterface interface.
- */
-interface TypeInterface
-{
-    public function getContent(): array;
-
-    public function setContent(array $array_data): void;
-}
