+++ src/Stack.php
