+++ src/Str.php
